requirements.txt
azure-functions
azure-cosmos
azure-ai-formrecognizer